test(server): singular graph routes
