+++ apps/server/test/graph/:graphId/[GET].test.ts
@@ -10,29 +10,29 @@ const dbName = oneUp(__filename)
 const db = getSource(dbName)
 let app: Express
 let api: Api
-let _delete: jest.SpyInstance
+let findOneById: jest.SpyInstance
 
 beforeAll(() =>
   setup(db).then(() => {
     api = new Api(db)
     app = getApp(api)
-    _delete = jest.spyOn(api.graphRepo, 'delete')
+    findOneById = jest.spyOn(api.graphRepo, 'findOneById')
   })
 )
 beforeEach(() => jest.clearAllMocks())
 afterAll(() => teardown(db))
 
 const testRequest = async () =>
-  request(app).delete('/graphs/58201858-5ce5-4ceb-8568-eecf55841b9f')
+  request(app).get('/graph/58201858-5ce5-4ceb-8568-eecf55841b9f')
 
-test('`_delete` is called once', async () => {
+test('`findOneById` is called once', async () => {
   await testRequest()
 
-  expect(_delete).toBeCalledTimes(1)
+  expect(findOneById).toBeCalledTimes(1)
 })
 
-test('`_delete` is called with correct args', async () => {
+test('`findOneById` is called with correct args', async () => {
   await testRequest()
 
-  expect(_delete).toBeCalledWith({ id: '58201858-5ce5-4ceb-8568-eecf55841b9f' })
+  expect(findOneById).toBeCalledWith('58201858-5ce5-4ceb-8568-eecf55841b9f')
 })
